minor refactor to make it easier to find policies related to manipulation
