+++ src/home_robot/home_robot/manipulation/__init__.py
@@ -1,0 +1,1 @@
+from .heuristic_place_policy import HeuristicPlacePolicy
