+++ src/home_robot/home_robot/manipulation/heuristic_place_policy.py
