--- src/home_robot/home_robot/place_policy/__init__.py
@@ -1,1 +1,0 @@
-from .heuristic_place_policy import HeuristicPlacePolicy
